fix orderField
