+++ Frontend/booking-sport/src/app/(page)/OrderField/BookingClientWrapper.tsx
@@ -1,0 +1,8 @@
+
+'use client';
+
+import BookingPage from './BookingPage';
+
+export default function BookingClientWrapper() {
+  return <BookingPage />;
+}
